feat(manager): enhanced UI with node management and database views
Major UI improvements:
- Add modal dialogs for node configuration
- Implement database listing per node
- Add status polling with connection state tracking
- Improve canvas-based node visualization
- Add Tailwind CSS for styling

Backend enhancements:
- Add databases API endpoint for listing databases
- Improve node status model with database info
- Add per-node HTTP client caching
- Implement poller lifecycle management

Build tooling:
- Add Justfile for development commands
- Add Tailwind CSS configuration
- Update Elm dependencies

+++ manager/src/api/instance.rs
@@ -7,7 +7,7 @@ use crate::state::AppState;
 /// Information about the local instance
 #[derive(Debug, Serialize)]
 pub struct LocalInstanceInfo {
-    pub port: u16,
+    pub port: u32,
     pub running: bool,
     pub url: String,
 }
